requirements.txt
azure-functions
azure-storage-blob
azure-identity
tensorflow
scikit-learn
pandas
numpy
allennlp
allennlp-models
nltk
